#11 代用品提案プロンプトの切り出し
